// src/features/source/pane/types.ts
export interface SelectionBox {
  startX: number
  startY: number
  currentX: number
  currentY: number
}